# docs/openapi.yaml
openapi: 3.1.0
info:
  title: IG-API - Instagram REST API
  description: REST API that integrates with Meta Instagram Graph API to retrieve posts and handle webhook events
  version: 1.0.0
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  contact:
    name: IG-API Support

tags:
  - name: Health
    description: Health check endpoints
  - name: Posts
    description: Instagram posts management
  - name: Webhooks
    description: Meta webhook endpoints

servers:
  - url: https://{api-id}.execute-api.{region}.amazonaws.com/{stage}
    description: Production (AWS API Gateway)
    variables:
      api-id:
        default: '0kp591p5k7'
        description: API Gateway ID
      region:
        default: eu-west-1
      stage:
        default: pre
        enum: [pre, int, pro]

components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
      description: API key for /posts endpoints
    WebhookSignature:
      type: apiKey
      in: header
      name: x-hub-signature-256
      description: HMAC-SHA256 signature for /webhooks POST

  schemas:
    Post:
      type: object
      properties:
        id:
          type: string
          description: Instagram post ID
        caption:
          type: string
          description: Post caption
        mediaType:
          type: string
          enum: [IMAGE, VIDEO, CAROUSEL_ALBUM]
        mediaUrl:
          type: string
          format: uri
          description: CDN URL for media
        permalink:
          type: string
          format: uri
          description: Instagram permalink
        timestamp:
          type: string
          format: date-time
        likeCount:
          type: integer
        commentsCount:
          type: integer
        thumbnailUrl:
          type: string
          format: uri
          description: URL de la miniatura (para videos)
        createdAt:
          type: string
          format: date-time
          description: Fecha de creación en DynamoDB
        lastVerificationDate:
          type: string
          format: date-time
          description: Last time post was verified against Meta API
        expiresAt:
          type: integer
          description: DynamoDB TTL timestamp

    Error:
      type: object
      properties:
        error:
          type: string
        stack:
          type: string
          description: Solo presente en modo desarrollo

    WebhookPayload:
      type: object
      properties:
        object:
          type: string
          enum: [instagram, page]
        entry:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              time:
                type: integer
              changes:
                type: array
                items:
                  type: object

paths:
  /health:
    get:
      summary: Health check
      operationId: getHealth
      tags: [Health]
      responses:
        '200':
          description: Service is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
                  timestamp:
                    type: string
                    format: date-time
              examples:
                healthCheck:
                  value:
                    status: ok
                    timestamp: '2026-06-07T10:00:00.000Z'
        '429':
          description: Too Many Requests - Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          headers:
            Retry-After:
              schema:
                type: integer
              description: Seconds to wait before retrying

  /posts:
    get:
      summary: List posts from Meta API
      operationId: listPosts
      tags: [Posts]
      security:
        - ApiKeyAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          description: Number of posts to fetch
      responses:
        '200':
          description: List of posts
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Post'
                  paging:
                    type: object
              examples:
                listPosts:
                  value:
                    data:
                      - id: '17841478291207902'
                        caption: 'Example post'
                        mediaType: IMAGE
                        mediaUrl: 'https://example.com/image.jpg'
                        permalink: 'https://instagram.com/p/abc123'
                        timestamp: '2026-06-07T10:00:00+0000'
                        likeCount: 42
                        commentsCount: 5
                        thumbnailUrl: null
                        createdAt: '2026-06-07T10:00:00.000Z'
                        expiresAt: 1725580800
                    paging:
                      cursors:
                        before: 'abc123'
                        after: 'def456'
        '400':
          description: Invalid limit parameter
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Missing API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Invalid API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Too Many Requests - Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          headers:
            Retry-After:
              schema:
                type: integer
              description: Seconds to wait before retrying
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '502':
          description: Bad Gateway - Meta API error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /posts/list:
    get:
      summary: List cached posts from DynamoDB
      description: Reads posts from DynamoDB with cursor-based pagination.
      tags: [Posts]
      security:
        - ApiKeyAuth: []
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          description: Number of posts to return (1-100)
        - in: query
          name: cursor
          schema:
            type: string
          description: Pagination cursor from previous response
      responses:
        '200':
          description: Paginated list of posts
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                  paging:
                    type: object
                    properties:
                      nextCursor:
                        type: string
                        nullable: true
        '400':
          description: Invalid limit parameter
        '401':
          description: Unauthorized

  /posts/{id}:
    get:
      summary: Get single post by ID
      operationId: getPost
      tags: [Posts]
      security:
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Instagram post ID
      responses:
        '200':
          description: Post details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        '400':
          description: Invalid post ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Missing or invalid API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Post not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Too Many Requests - Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          headers:
            Retry-After:
              schema:
                type: integer
              description: Seconds to wait before retrying
        '502':
          description: Bad Gateway - Meta API error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /posts/sync:
    post:
      summary: Sync posts from Meta API to DynamoDB
      operationId: syncPosts
      tags: [Posts]
      security:
        - ApiKeyAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                limit:
                  type: integer
                  minimum: 1
                  maximum: 100
                  default: 20
      responses:
        '200':
          description: Sync completed
          content:
            application/json:
              schema:
                type: object
                properties:
                  synced:
                    type: integer
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Post'
        '400':
          description: Invalid limit
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Missing or invalid API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Too Many Requests - Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          headers:
            Retry-After:
              schema:
                type: integer
              description: Seconds to wait before retrying
        '502':
          description: Bad Gateway - Meta API error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /posts/verify:
    post:
      summary: Verify stale posts against Meta API
      operationId: verifyPosts
      tags: [Posts]
      security:
        - ApiKeyAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                limit:
                  type: integer
                  minimum: 1
                  maximum: 100
                  default: 50
      responses:
        '200':
          description: Verification completed
          content:
            application/json:
              schema:
                type: object
                properties:
                  summary:
                    type: object
                    properties:
                      total:
                        type: integer
                      verified:
                        type: integer
                      deleted:
                        type: integer
                      rateLimited:
                        type: integer
                      errors:
                        type: integer
                  results:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                        status:
                          type: string
                          enum: [verified, deleted, rate_limited, error]
                        error:
                          type: string
        '400':
          description: Invalid limit
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Missing or invalid API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Too Many Requests - Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          headers:
            Retry-After:
              schema:
                type: integer
              description: Seconds to wait before retrying
        '502':
          description: Bad Gateway - Meta API error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /webhooks:
    get:
      summary: Webhook subscription verification (Meta challenge-response)
      operationId: verifyWebhookSubscription
      tags: [Webhooks]
      parameters:
        - name: hub.mode
          in: query
          required: true
          schema:
            type: string
            enum: [subscribe]
        - name: hub.challenge
          in: query
          required: true
          schema:
            type: string
        - name: hub.verify_token
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Challenge returned
          content:
            text/plain:
              schema:
                type: string
        '403':
          description: Invalid mode or token
          content:
            text/plain:
              schema:
                type: string
        '429':
          description: Too Many Requests - Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          headers:
            Retry-After:
              schema:
                type: integer
              description: Seconds to wait before retrying

    post:
      summary: Receive webhook events from Meta
      operationId: receiveWebhookEvent
      tags: [Webhooks]
      security:
        - WebhookSignature: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WebhookPayload'
      responses:
        '200':
          description: Event received (always returns 200 to Meta)
          content:
            text/plain:
              schema:
                type: string
                example: EVENT_RECEIVED
        '401':
          description: Invalid signature
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Too Many Requests - Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          headers:
            Retry-After:
              schema:
                type: integer
              description: Seconds to wait before retrying
        '500':
          description: Internal Server Error - META_APP_SECRET not configured
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'